feat: PERAR-26255 PERAR-26466 Convertion BigDecimal amount

+++ artifact/dtos/RBVDC406/src/main/java/com/bbva/rbvd/dto/insuranceenterprise/commons/dto/AmountDTO.java
@@ -7,14 +7,15 @@ public class AmountDTO implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private BigDecimal amount;
+    private Double amount;
     private String currency;
 
-    public BigDecimal getAmount() {
+
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
